add by mass

+++ src/main/java/com/tdsi/sn/app_moblile_api/Entity/Etudiant.java
@@ -23,5 +23,5 @@ public class Etudiant {
     private int motPasse ;
     private  int solde ;
     private String numero_carte ;
-    private BigInteger pu;
+    private int pu;
 }
